refactor(skills): split diagnose/index.ts into modules
Decomposes the 1,264-line monolithic src/skills/diagnose/index.ts into
four cohesive modules under a new _phases/ directory, mirroring the
convention already used by the sibling skill src/skills/mint/_phases/.

New layout:
  src/skills/diagnose/
    index.ts                  (~60 LOC) — thin orchestrator; re-exports public surface
    _phases/
      types.ts               (~155 LOC) — Zod schemas + inferred types
      triage.ts              (~110 LOC) — classifyAndExtract, fixSpansMultipleFiles, computeOutcome
      verifier.ts            (~230 LOC) — parseShadowVerifyOutput, runReproducerBaseline, buildVerifierUserPrompt, autoVerifyHypotheses
      orchestrator.ts        (~360 LOC) — handler + worktree testing + tool-permission helpers

Pure structural move — no logic, string, or behavior changes.

index.ts re-exports all previously public symbols so no external import
path changes: diagnose.test.ts and diagnose-auto-verify.test.ts import
from 'skills/diagnose/index.js' and continue to work unchanged.

Prompt path resolution is unaffected: loadSkillPrompts('diagnose') is
called in orchestrator.ts and resolves via prompt-loader.ts's own
import.meta.url anchor (src/skills/_lib/) — two hops up to src/, then
descends to src/skills/diagnose/prompts/. Moving code to _phases/ does
not change this resolution path.

Verification:
- pnpm lint (tsc --noEmit): PASS (zero errors)
- pnpm test -- src/skills/diagnose: 115/115 tests PASS
  (src/skills/diagnose.test.ts: 95 tests; src/skills/diagnose-auto-verify.test.ts: 20 tests)
